+++ python/ray/data/tests/test_arrow_block_scaling.py
@@ -1,13 +1,48 @@
 import gc
+import os
+import sys
+from tempfile import TemporaryDirectory
 
+import pyarrow as pa
 import pytest
+from pyarrow import parquet as pq
 
 import ray
 from ray.data import DataContext
-from ray.data._internal.util import GiB
-from ray.data.tests.test_arrow_block import (
-    parquet_dataset_single_column_gt_2gb,  # noqa: F401
-)
+from ray.data._internal.util import GiB, MiB
+
+
+@pytest.fixture(scope="module")
+def parquet_dataset_single_column_gt_2gb():
+    chunk_size = 256 * MiB
+    num_chunks = 10
+
+    total_column_size = chunk_size * 10  # ~2.5 GiB
+
+    with TemporaryDirectory() as tmp_dir:
+        dataset_path = f"{tmp_dir}/large_parquet_chunk_{chunk_size}"
+
+        # Create directory
+        os.mkdir(dataset_path)
+
+        for i in range(num_chunks):
+            chunk = b"a" * chunk_size
+
+            d = {"id": [i], "bin": [chunk]}
+            t = pa.Table.from_pydict(d)
+
+            print(f">>> Table schema: {t.schema} (size={sys.getsizeof(t)})")
+
+            filepath = f"{dataset_path}/chunk_{i}.parquet"
+            pq.write_table(t, filepath)
+
+            print(f">>> Created a chunk #{i}")
+
+        print(f">>> Created dataset at {dataset_path}")
+
+        yield dataset_path, num_chunks, total_column_size
+
+        print(f">>> Cleaning up dataset at {dataset_path}")
 
 
 @pytest.mark.parametrize(
@@ -19,7 +54,7 @@ from ray.data.tests.test_arrow_block import (
 )
 def test_arrow_batch_gt_2gb(
     ray_start_regular,
-    parquet_dataset_single_column_gt_2gb,  # noqa: F811
+    parquet_dataset_single_column_gt_2gb,
     restore_data_context,
     op,
 ):
@@ -67,3 +102,7 @@ def test_arrow_batch_gt_2gb(
     del ds
     # Force GC to free up object store memory
     gc.collect()
+
+
+if __name__ == "__main__":
+    sys.exit(pytest.main(["-v", __file__]))
